+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## v0.1.2
-- Added support for installing Python on different linux distros
+- Added support for installing Python on most Linux distros
 - Fixed a bug related to creating `pyflow` directory
 
 ## v0.1.1